Added SqlServer InfoMessage logs.
The static SqlHelper.LogSqlServerInfoMessage drives whether
info messages will be logged or not.
The event is always registered. This should not impact the performance
but if it's the case, SqlConnectionControllers should subscribe/unsubscribe
to the InfoMessage event dynamically. There are 2 options:
- An internal static OnChange event or a dynamic concurrent
bag of SqlConnectionController should be managed... (costly)
- A check of a private _isLogging vs. the static before (and after) each call
to the db (more efficient). Unfortunately, the ISqlCommandExecutor is
totally independent of the CallContext/ConnectionController (and this
is fine).

==> for the moment, the event is always subscribed.